+++ R/delta.R
@@ -22,7 +22,7 @@ az_delta_files <- function(conn, url) {
       value = url
     )
   }
-  check_azure_url(url)
+  url <- check_azure_url(url)
   ensure_azure_exts(conn, delta = TRUE)
   try_as_tibble(az_query(
     conn,